+++ src/options/Options.ts
@@ -192,6 +192,12 @@ export class Options implements IOptions {
      * @type {boolean}
      */
     @IsBoolean()
+    public readonly shuffleStringArray!: boolean;
+
+    /**
+     * @type {boolean}
+     */
+    @IsBoolean()
     public readonly sourceMap!: boolean;
 
     /**