+++ test/dev/dev.ts
@@ -11,11 +11,15 @@ import { NO_ADDITIONAL_NODES_PRESET } from '../../src/options/presets/NoCustomNo
             var bar = 'bar';
             var baz = 'baz';
             var bark = 'bark';
+            
+            console.log(foo, bar, baz, bark);
         `,
         {
             ...NO_ADDITIONAL_NODES_PRESET,
             stringArray: true,
-            stringArrayThreshold: 1
+            stringArrayThreshold: 1,
+            rotateStringArray: true,
+            shuffleStringArray: true
         }
     ).getObfuscatedCode();